add userUsecases files for atomic db operations
