+++ src/usecases/userUsecases/createSingleUser.js
@@ -1,0 +1,3 @@
+const createSingleUser = async (user) => await User.create(user)
+
+export { createSingleUser }
