+++ src/usecases/userUsecases/deleteSingleUser.js
@@ -1,0 +1,3 @@
+const deleteSingleUser = async (id) => await User.findByIdAndDelete(id)
+
+export { deleteSingleUser }
