+++ src/usecases/userUsecases/getAllUsers.js
@@ -1,0 +1,3 @@
+const getAllUsers = async () => await User.find()
+
+export { getAllUsers }
