+++ src/usecases/userUsecases/getLikesMadeByUser.js
@@ -1,0 +1,3 @@
+const getLikesMadeByUser = async (id) => await User.findById(id)
+
+export { getLikesMadeByUser }
