+++ src/usecases/userUsecases/getPlacesCreatedByUser.js
@@ -1,0 +1,3 @@
+const getPlacesCreatedByUser = async (id) => await User.findById(id)
+
+export { getPlacesCreatedByUser }
