+++ src/usecases/userUsecases/getReviewsMadeByUser.js
@@ -1,0 +1,3 @@
+const getReviewsMadeByUser = async (id) => await User.findById(id)
+
+export { getReviewsMadeByUser }
