+++ src/usecases/userUsecases/getRoutesCreatedByUser.js
@@ -1,0 +1,3 @@
+const getRoutesCreatedByUser = async (id) => await User.findById(id)
+
+export { getRoutesCreatedByUser }
