+++ src/usecases/userUsecases/getSingleUser.js
@@ -1,0 +1,3 @@
+const getSingleUser = async (id) => await User.findById(id)
+
+export { getSingleUser }
