+++ src/usecases/userUsecases/updateSingleUser.js
@@ -1,0 +1,4 @@
+const updateSingleUser = async (id, body) =>
+  await User.findByIdAndUpdate(id, body)
+
+export { updateSingleUser }
